refactor: move enametoolong
